backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0

pydantic-settings>=2.2,<3.0

SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
pgvector>=0.2,<1.0

alembic>=1.13,<2.0

python-multipart>=0.0.9,<1.0
pypdf>=4.0,<6.0

openai>=1.0,<2.0
python-dotenv>=1.0,<2.0

python-jose[cryptography]>=3.3,<4.0

boto3>=1.30,<2.0

auth0-fastapi-api
